Overhauls and updates library automation (#11)
- You can now re-build from Github Actions UI
- uw-saml-poetry will now have its python dependencies patched (instead of just the OS)
- When uw-saml-poetry python dependencies are patched, the workflow will open a Pull Request to be approved
  by @tomthorogood
- When the pull request is merged, a separate workflow will extract the build-id from the PR and
  tag the new builds as "latest"
- You can now promote any tag to `latest` from the Github Actions UI, if the tag exists for a given image.

**Note:**

OS-only updates (i.e., no changes to `poetry.lock`) will not generate any PR, so there will be nothing
to automatically "merge". Therefore, automated builds that only patch the OS will not automatically be promoted to "latest". To promote them, you must manually trigger the `tag-latest` workflow.

We can address an extra level of automation in the future if desired or required.